+++ src/main/core/blow/exception/PluginAbortException.groovy
@@ -17,30 +17,14 @@
  *   along with Blow.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package blow.util
+package blow.exception
 
+import groovy.transform.InheritConstructors
 
 /**
  *
  *  @author Paolo Di Tommaso <paolo.ditommaso@gmail.com>
  */
-
-import ch.qos.logback.classic.Level
-import ch.qos.logback.classic.spi.ILoggingEvent
-import ch.qos.logback.core.CoreConstants
-import ch.qos.logback.core.LayoutBase
-
-class PrettyConsoleLayout extends LayoutBase<ILoggingEvent> {
-
-    public String doLayout(ILoggingEvent event) {
-        StringBuilder buffer = new StringBuilder(128);
-        if( event.getLevel() != Level.INFO ) {
-            buffer.append( event.getLevel().toString() ) .append(": ")
-        }
-
-        return buffer
-                .append(event.getFormattedMessage())
-                .append(CoreConstants.LINE_SEPARATOR)
-                .toString()
-    }
+@InheritConstructors
+class PluginAbortException extends BlowException {
 }