+++ src/main/java/blow/eventbus/HandlerFindingStrategy.java
@@ -1,8 +1,30 @@
+/*
+ * Copyright (c) 2012, the authors.
+ *
+ *   This file is part of Blow.
+ *
+ *   Blow is free software: you can redistribute it and/or modify
+ *   it under the terms of the GNU General Public License as published by
+ *   the Free Software Foundation, either version 3 of the License, or
+ *   (at your option) any later version.
+ *
+ *   Blow is distributed in the hope that it will be useful,
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *   GNU General Public License for more details.
+ *
+ *   You should have received a copy of the GNU General Public License
+ *   along with Blow.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 package blow.eventbus;
 
 import com.google.common.collect.Multimap;
 
 /**
+ * NOTE THIS FILE IS A DERIVATIVE WORK OF THE GOOGLE GUAVA LIBRARY
+ *
+ *
  * A method for finding event handler methods in objects, for use by
  * {@link EventBus}.
  *